+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [CSS module for Sidebar component](styles/Sidebar.module.css)
 - Path variable `@` to [TypeScript configuration](tsconfig.json)
 - [Type declaration](db.d.ts) for user defined links
+- [Header component](components/Header.tsx)
 
 ### Changed